requirements.txt
bcrypt==3.2.2
fastapi==0.78.0
passlib==1.7.4
pymongo==4.1.1
python-jose==3.3.0
python-multipart==0.0.5
uvicorn==0.18.2
Pillow==9.1.1
python-decouple==3.6
